Vertex and Fragment shaders now load from external files (with flag to allow for STRINGIFY code to be loaded instead). This was done to pull the IP (shader code) out of the public git repo.
The files (.glslf and .glslv) must be present on the machine and therefore this is not a good mechanism for deploying compiled plugins (the files load dynamically at runtime).

The benefit of this setup is when debugging, no longer have to recompile the plugin for Shader change. Just reopen the plugin in whatever software is being used (after changing the .glsl* files).